tanmatra

Module nucleus

Nuclear structure: binding energy, mass defect, nuclear radius.

Uses the Bethe-Weizsacker semi-empirical mass formula (liquid drop model) with standard coefficients.

Structs

Nucleus

A nucleus characterized by atomic number Z and mass number A.

Constants

CARBON12

Carbon-12.

HELIUM4

Helium-4 (alpha particle).

HYDROGEN

Hydrogen-1 (protium).

IRON56

Iron-56 (peak of binding energy curve).

OXYGEN16

Oxygen-16.

URANIUM235

Uranium-235 (fissile).

URANIUM238

Uranium-238 (most common uranium isotope).

Functions

binding_energy_mev

Computes the nuclear binding energy (in MeV) using the Bethe-Weizsacker semi-empirical mass formula.

binding_energy_per_nucleon

Binding energy per nucleon (in MeV).

is_magic_number

Checks whether a number is a nuclear magic number.

mass_defect_mev

Mass defect (in MeV): equals the binding energy in the semi-empirical model.

nuclear_radius_fm

Nuclear radius (in fm) using the empirical formula R = R0 * A^(1/3).